did a lot of things, added mobx support for the memory component, renamed a few things as well
Okay so i added mobx support to the memory component and now its state is managed by mobx. I also renamed the video component to MemoryVideoComponent because the name better explain the purpose of the component and the app as a whole. I also did the same thing with the video service (renamed to RandomMemoryStoreService)