Merge Feature/setup mobx into master (#4)
Okay so i added mobx support to the memory component and now its state is managed by mobx. I also renamed the video component to MemoryVideoComponent because the name better explain the purpose of the component and the app as a whole. I also did the same thing with the video service (renamed to RandomMemoryStoreService)

I also fixed the uppercase on the LoginComponent file name

Co-authored-by: Puma <c.coque@hotmail.fr>